o/snapstate: move prerequisite handler and related functions into new file
